feat: Add affected blocklist to data and revamp dashboard to use it
This innocuous seeming change was actually a massive time sink. The dashboard is way more dynamic now and has more data.

We have two charts now, so I think it's better that the annotations are organization wide instead of repeating the same annotation twice and hardcoding the IDs for the charts.

Closes: #3
Closes: #4